requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
face_recognition==1.3.0
sqlalchemy==2.0.32
pydantic==2.7.0
python-dotenv==0.19.2
bcrypt==3.2.2
passlib==1.7.4